Files are now split according to topic.
